Update generators for control blocks
